@@ -3,28 +3,27 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@craco/craco": "^6.4.0",
-    "@reduxjs/toolkit": "^1.6.2",
-    "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^11.1.0",
-    "@testing-library/user-event": "^12.1.10",
-    "@types/lodash": "^4.14.176",
+    "@reduxjs/toolkit": "^1.9.0",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^13.4.0",
+    "@testing-library/user-event": "^14.4.3",
+    "@types/lodash": "^4.14.189",
     "lodash": "^4.17.21",
     "lodash.uniq": "^4.5.0",
-    "prettier": "^2.5.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-redux": "^7.2.6",
-    "react-scripts": "4.0.3",
-    "react-select": "^5.5.0",
-    "sass": "^1.51.0",
+    "prettier": "^2.7.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-redux": "^8.0.5",
+    "react-scripts": "5.0.1",
+    "react-select": "^5.6.1",
+    "sass": "^1.56.1",
     "styled-components": "^5.3.6",
-    "web-vitals": "^1.0.1"
+    "web-vitals": "^3.1.0"
   },
   "scripts": {
-    "start": "craco start",
-    "build": "craco build",
-    "test": "craco test --collectCoverage --watchAll",
+    "start": "react-scripts start",
+    "build": "react-scripts build",
+    "test": "react-scripts --collectCoverage --watchAll",
     "eject": "react-scripts eject",
     "prettier": "prettier --write \"**/*.{tsx,ts,js,css}\"",
     "fast": "git add . && git commit --amend && git push --force-with-lease"
@@ -48,13 +47,13 @@
     ]
   },
   "devDependencies": {
-    "@types/react": "^17.0.30",
+    "@types/react": "^18.0.25",
     "@types/styled-components": "^5.1.26",
-    "autoprefixer": "^9.8.8",
+    "autoprefixer": "^10.4.13",
     "file-loader": "^6.2.0",
-    "postcss": "^7.0.39",
-    "react-router-dom": "^6.0.0",
+    "postcss": "^8.4.19",
+    "react-router-dom": "^6.4.3",
     "tailwindcss": "npm:@tailwindcss/postcss7-compat@^2.2.17",
-    "typescript": "^4.4.4"
+    "typescript": "^4.9.3"
   }
 }
